🔨 Refactor(#79): sectionsApis 지역성 원칙 준수
