+++ src/features/learn/apis.ts
@@ -1,5 +1,5 @@
-import Section from '@type/Section';
 import api from '@/axios/instance';
+import type { Section } from '@/features/learn/types';
 
 const sectionsApis = {
   // 특정 섹션 ID에 대한 데이터 가져오기
