+++ src/features/learn/types.ts
@@ -1,0 +1,11 @@
+export interface Part {
+  id: number;
+  sectionId: number;
+  name: string;
+}
+
+export interface Section {
+  id: number;
+  name: string;
+  part: Part[];
+}
