--- src/types/Part.ts
@@ -1,5 +1,0 @@
-export default interface Part {
-  id: number;
-  sectionId: number;
-  name: string;
-}
