--- src/types/Section.ts
@@ -1,7 +1,0 @@
-import Part from './Part';
-
-export default interface Section {
-  id: number;
-  name: string;
-  part: Part[];
-}
